+++ include/coordinates.h
@@ -2,10 +2,25 @@
 #ifndef _COORDINATES_H
 #define _COORDINATES_H
 
-#include <iostream>
-#include "geom.h"
+// Lightweight types and functions for coordinates, for classes that don't
+// need to pull in boost::geometry.
+//
+// Things that pull in boost::geometry should go in coordinates_geom.h
+
 #include <utility>
+#include <vector>
+#include <deque>
 #include <unordered_set>
+
+// A 36-bit integer can store all OSM node IDs; we represent this as 16 collections
+// of 32-bit integers.
+#define NODE_SHARDS 16
+typedef uint32_t ShardedNodeID;
+typedef uint64_t NodeID;
+typedef uint64_t WayID;
+
+typedef std::vector<WayID> WayVec;
+
 
 #ifdef FAT_TILE_INDEX
 typedef uint32_t TileCoordinate;
@@ -101,35 +116,8 @@ double degp2meter(double degp, double latp);
 
 double meter2degp(double meter, double latp);
 
-void insertIntermediateTiles(Linestring const &points, uint baseZoom, std::unordered_set<TileCoordinates> &tileSet);
-void insertIntermediateTiles(Ring const &points, uint baseZoom, std::unordered_set<TileCoordinates> &tileSet);
-
 // the range between smallest y and largest y is filled, for each x
 void fillCoveredTiles(std::unordered_set<TileCoordinates> &tileSet);
-
-// ------------------------------------------------------
-// Helper class for dealing with spherical Mercator tiles
-
-class TileBbox { 
-
-public:
-	double minLon, maxLon, minLat, maxLat, minLatp, maxLatp;
-	double xmargin, ymargin, xscale, yscale;
-	TileCoordinates index;
-	uint zoom;
-	bool hires;
-	bool endZoom;
-	Box clippingBox;
-
-	TileBbox(TileCoordinates i, uint z, bool h, bool e);
-
-	std::pair<int,int> scaleLatpLon(double latp, double lon) const;
-	MultiPolygon scaleGeometry(MultiPolygon const &src) const;
-	std::pair<double, double> floorLatpLon(double latp, double lon) const;
-
-	Box getTileBox() const;
-	Box getExtendBox() const;
-};
 
 #endif //_COORDINATES_H
