+++ src/osm_lua_processing.cpp
@@ -1,6 +1,8 @@
+#include <iostream>
+
 #include "osm_lua_processing.h"
 #include "helpers.h"
-#include <iostream>
+#include "coordinates_geom.h"
 
 
 using namespace std;
