+++ src/output_object.cpp
@@ -4,6 +4,7 @@
 
 #include "output_object.h"
 #include "helpers.h"
+#include "coordinates_geom.h"
 #include <iostream>
 using namespace std;
 namespace geom = boost::geometry;
